refactor: de-coupled e2e tests from config.ts
