+++ packages/cip30/package.json
@@ -32,6 +32,7 @@
     "lint:fix": "yarn lint --fix",
     "test": "jest -c ./jest.config.js",
     "test:build:verify": "tsc --build ./test",
+    "test:e2e": "shx echo 'test:e2e' command not implemented yet",
     "coverage": "shx echo No coverage report for this package",
     "prepack": "yarn build",
     "test:debug": "DEBUG=true yarn test"
