+++ packages/e2e/src/factories.ts
@@ -9,8 +9,7 @@ import {
   RewardsProvider,
   StakePoolProvider,
   TxSubmitProvider,
-  UtxoProvider,
-  WalletProvider
+  UtxoProvider
 } from '@cardano-sdk/core';
 import {
   BlockFrostAPI,
@@ -19,16 +18,18 @@ import {
   blockfrostNetworkInfoProvider,
   blockfrostRewardsProvider,
   blockfrostTxSubmitProvider,
-  blockfrostUtxoProvider,
-  blockfrostWalletProvider
+  blockfrostUtxoProvider
 } from '@cardano-sdk/blockfrost';
 import { CardanoWalletFaucetProvider, FaucetProvider } from './FaucetProvider';
 import { KeyManagement } from '@cardano-sdk/wallet';
+import { LogLevel, createLogger } from 'bunyan';
+import { Logger } from 'ts-log';
 import { createConnectionObject } from '@cardano-ogmios/client';
 import { createStubStakePoolProvider } from '@cardano-sdk/util-dev';
 import { ogmiosTxSubmitProvider } from '@cardano-sdk/ogmios';
 import { txSubmitHttpProvider } from '@cardano-sdk/cardano-services-client';
 import DeviceConnection from '@cardano-foundation/ledgerjs-hw-app-cardano';
+import memoize from 'lodash/memoize';
 
 const BLOCKFROST_PROVIDER = 'blockfrost';
 const BLOCKFROST_MISSING_PROJECT_ID = 'Missing project id';
@@ -61,8 +62,8 @@ export const networkInfoProviderFactory = new ProviderFactory<NetworkInfoProvide
 export const rewardsProviderFactory = new ProviderFactory<RewardsProvider>();
 export const txSubmitProviderFactory = new ProviderFactory<TxSubmitProvider>();
 export const utxoProviderFactory = new ProviderFactory<UtxoProvider>();
-export const walletProviderFactory = new ProviderFactory<WalletProvider>();
 export const stakePoolProviderFactory = new ProviderFactory<StakePoolProvider>();
+export const loggerFactory = new ProviderFactory<Logger>();
 
 // Faucet providers
 faucetProviderFactory.register('cardano-wallet', CardanoWalletFaucetProvider.create);
@@ -175,19 +176,6 @@ utxoProviderFactory.register(
     })
 );
 
-// Wallet providers
-walletProviderFactory.register(
-  blockfrostWalletProvider.name,
-  async (params: any): Promise<WalletProvider> =>
-    new Promise<WalletProvider>(async (resolve) => {
-      if (params.isTestnet === undefined) throw new Error(BLOCKFROST_MISSING_FLAG);
-
-      if (params.projectId === undefined) throw new Error(BLOCKFROST_MISSING_PROJECT_ID);
-
-      resolve(blockfrostWalletProvider(await getBlockfrostApi(params.isTestnet, params.projectId)));
-    })
-);
-
 // Stake Pool providers
 stakePoolProviderFactory.register(
   'stub',
@@ -256,3 +244,31 @@ keyManagementFactory.register('trezor', async (params: any): Promise<KeyManageme
     })
   );
 });
+
+/**
+ * Utility function to create key agents at different account indices.
+ *
+ * @param accountIndex The ccount index.
+ * @param provider The provider.
+ * @param params The provider parameters.
+ * @returns a key agent.
+ */
+export const keyAgentById = memoize(async (accountIndex: number, provider: string, params: any) => {
+  params.accountIndex = accountIndex;
+  return keyManagementFactory.create(provider, params);
+});
+
+// Logger
+
+/**
+ * Gets the logger instance.
+ *
+ * @param severity The minimum severity of the log messages that will be logged.
+ * @returns The Logger instance.
+ */
+export const getLogger = function (severity: string): Logger {
+  return createLogger({
+    level: severity as LogLevel,
+    name: 'e2e tests'
+  });
+};
