+++ packages/e2e/test/wallet/SingleAddressWallet/delegation.test.ts
@@ -4,19 +4,21 @@ import { Cardano } from '@cardano-sdk/core';
 import { ObservableWallet, SingleAddressWallet, StakeKeyStatus, Transaction } from '@cardano-sdk/wallet';
 import { TX_TIMEOUT, firstValueFromTimed, waitForWalletStateSettle } from '../util';
 import {
-  assetProvider,
-  chainHistoryProvider,
+  assetProviderFactory,
+  chainHistoryProviderFactory,
   keyAgentById,
-  networkInfoProvider,
-  poolId1,
-  poolId2,
-  rewardsProvider,
-  stakePoolProvider,
-  txSubmitProvider,
-  utxoProvider,
-  walletProvider
-} from '../../config';
+  networkInfoProviderFactory,
+  rewardsProviderFactory,
+  stakePoolProviderFactory,
+  txSubmitProviderFactory,
+  utxoProviderFactory
+} from '../../../src/factories';
+import { env } from '../environment';
+
 import { combineLatest, filter, firstValueFrom } from 'rxjs';
+
+export const poolId1 = Cardano.PoolId(env.POOL_ID_1);
+export const poolId2 = Cardano.PoolId(env.POOL_ID_2);
 
 const getWalletStateSnapshot = async (wallet: ObservableWallet) => {
   const [rewardAccount] = await firstValueFrom(wallet.delegation.rewardAccounts$);
@@ -77,15 +79,20 @@ const getWallet = async (idx: number) =>
   new SingleAddressWallet(
     { name: `Test Wallet ${idx}` },
     {
-      assetProvider: await assetProvider,
-      chainHistoryProvider: await chainHistoryProvider,
-      keyAgent: await keyAgentById(idx),
-      networkInfoProvider: await networkInfoProvider,
-      rewardsProvider: await rewardsProvider,
-      stakePoolProvider: await stakePoolProvider,
-      txSubmitProvider: await txSubmitProvider,
-      utxoProvider: await utxoProvider,
-      walletProvider: await walletProvider
+      assetProvider: await assetProviderFactory.create(env.ASSET_PROVIDER, env.ASSET_PROVIDER_PARAMS),
+      chainHistoryProvider: await chainHistoryProviderFactory.create(
+        env.CHAIN_HISTORY_PROVIDER,
+        env.CHAIN_HISTORY_PROVIDER_PARAMS
+      ),
+      keyAgent: await keyAgentById(idx, env.KEY_MANAGEMENT_PROVIDER, env.KEY_MANAGEMENT_PARAMS),
+      networkInfoProvider: await networkInfoProviderFactory.create(
+        env.NETWORK_INFO_PROVIDER,
+        env.NETWORK_INFO_PROVIDER_PARAMS
+      ),
+      rewardsProvider: await rewardsProviderFactory.create(env.REWARDS_PROVIDER, env.REWARDS_PROVIDER_PARAMS),
+      stakePoolProvider: await stakePoolProviderFactory.create(env.STAKE_POOL_PROVIDER, env.STAKE_POOL_PROVIDER_PARAMS),
+      txSubmitProvider: await txSubmitProviderFactory.create(env.TX_SUBMIT_PROVIDER, env.TX_SUBMIT_PROVIDER_PARAMS),
+      utxoProvider: await utxoProviderFactory.create(env.UTXO_PROVIDER, env.UTXO_PROVIDER_PARAMS)
     }
   );
