+++ packages/e2e/test/wallet/SingleAddressWallet/pouchdbWalletStores.test.ts
@@ -1,15 +1,15 @@
 import { SingleAddressWallet, storage } from '@cardano-sdk/wallet';
 import {
-  assetProvider,
-  chainHistoryProvider,
-  keyAgent,
-  networkInfoProvider,
-  rewardsProvider,
-  stakePoolProvider,
-  txSubmitProvider,
-  utxoProvider,
-  walletProvider
-} from '../../config';
+  assetProviderFactory,
+  chainHistoryProviderFactory,
+  keyAgentById,
+  networkInfoProviderFactory,
+  rewardsProviderFactory,
+  stakePoolProviderFactory,
+  txSubmitProviderFactory,
+  utxoProviderFactory
+} from '../../../src/factories';
+import { env } from '../environment';
 import { filter, firstValueFrom } from 'rxjs';
 import { waitForWalletStateSettle } from '../util';
 import delay from 'delay';
@@ -26,16 +26,24 @@ describe('SingleAddressWallet/pouchdbWalletStores', () => {
     new SingleAddressWallet(
       { name: walletName },
       {
-        assetProvider: await assetProvider,
-        chainHistoryProvider: await chainHistoryProvider,
-        keyAgent: await keyAgent,
-        networkInfoProvider: await networkInfoProvider,
-        rewardsProvider: await rewardsProvider,
-        stakePoolProvider: await stakePoolProvider,
+        assetProvider: await assetProviderFactory.create(env.ASSET_PROVIDER, env.ASSET_PROVIDER_PARAMS),
+        chainHistoryProvider: await chainHistoryProviderFactory.create(
+          env.CHAIN_HISTORY_PROVIDER,
+          env.CHAIN_HISTORY_PROVIDER_PARAMS
+        ),
+        keyAgent: await keyAgentById(0, env.KEY_MANAGEMENT_PROVIDER, env.KEY_MANAGEMENT_PARAMS),
+        networkInfoProvider: await networkInfoProviderFactory.create(
+          env.NETWORK_INFO_PROVIDER,
+          env.NETWORK_INFO_PROVIDER_PARAMS
+        ),
+        rewardsProvider: await rewardsProviderFactory.create(env.REWARDS_PROVIDER, env.REWARDS_PROVIDER_PARAMS),
+        stakePoolProvider: await stakePoolProviderFactory.create(
+          env.STAKE_POOL_PROVIDER,
+          env.STAKE_POOL_PROVIDER_PARAMS
+        ),
         stores,
-        txSubmitProvider: await txSubmitProvider,
-        utxoProvider: await utxoProvider,
-        walletProvider: await walletProvider
+        txSubmitProvider: await txSubmitProviderFactory.create(env.TX_SUBMIT_PROVIDER, env.TX_SUBMIT_PROVIDER_PARAMS),
+        utxoProvider: await utxoProviderFactory.create(env.UTXO_PROVIDER, env.UTXO_PROVIDER_PARAMS)
       }
     );
