+++ packages/golden-test-generator/package.json
@@ -37,6 +37,7 @@
     "pretest": "yarn build",
     "test": "jest -c ./jest.config.js",
     "test:build:verify": "tsc --build ./test",
+    "test:e2e": "shx echo 'test:e2e' command not implemented yet",
     "coverage": "shx echo No coverage report for this package"
   },
   "dependencies": {
